refactor: extract skill availability ownership
